add compiler warnings and release optimization

+++ src/library/services/graphics/opengl/GLSLShader.cpp
@@ -179,9 +179,9 @@ void GLSLShader::cleanUp() {
 }
 
 int GLSLShader::getUniformLocation(const std::string &uniformName) {
-    int location = glGetUniformLocation(m_shaderProgram, uniformName.c_str());
+    GLint location = glGetUniformLocation(m_shaderProgram, uniformName.c_str());
 
-    if (location == GL_INVALID_INDEX) {
+    if ((unsigned) location == GL_INVALID_INDEX) {
         Logger::warn("Uniform " + uniformName + " not found in shader!");
     }
